+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Verify keepalive is used by clients by rejecting session reuse on new sockets,
   except in case where no-response mode is active
-
+* Add `mock-server-session-allow-missing-user-agent` session header to override
+  missing User-Agent check. 
 
 ## [1.11.0] - 2023-07-06
 ### Added